Use the menu in components js_only
