+++ assets/src/domain/eventEditor/ui/tickets/ticketCard/EditTicketButton.tsx
@@ -1,11 +1,11 @@
-import React, { CSSProperties } from 'react';
+import React from 'react';
 
 import { EspressoButton } from '@application/ui/input';
 import { EditItemButtonProps } from '../../../interfaces/types';
 import { useTicketContext } from '../../../hooks';
 import { useEditorModal } from '../../../../../application/ui/layout/editorModal';
 
-const EditTicketButton: React.FC<EditItemButtonProps> = ({ position }) => {
+const EditTicketButton: React.FC<EditItemButtonProps> = () => {
 	const { id: entityId } = useTicketContext();
 
 	const { openEditor } = useEditorModal();
@@ -16,21 +16,7 @@ const EditTicketButton: React.FC<EditItemButtonProps> = ({ position }) => {
 			entityId,
 		});
 
-	const style: CSSProperties = {
-		position: 'absolute',
-		right: '.5rem',
-		textAlign: 'right',
-		zIndex: 1,
-		...(position === 'top' && {
-			top: '.5rem',
-		}),
-	};
-
-	return (
-		<div style={style}>
-			<EspressoButton icon='edit' onClick={onClick} />
-		</div>
-	);
+	return <EspressoButton icon='edit' onClick={onClick} />;
 };
 
 export default EditTicketButton;
