+++ assets/src/domain/eventEditor/ui/tickets/ticketPriceCalculator/buttons/TicketPriceCalculatorButton.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { EspressoButton } from '@application/ui/input';
 
-import { TpcButtonDataProps, TpcModalProps } from '../types';
+import { TpcButtonDataProps } from '../types';
 import { useEditorModal } from '../../../../../../application/ui/layout/editorModal';
 
-const TicketPriceCalculatorButton: React.FC<TpcButtonDataProps & TpcModalProps> = ({ ticketId }) => {
+const TicketPriceCalculatorButton: React.FC<TpcButtonDataProps> = ({ ticketId }) => {
 	const { openEditor } = useEditorModal();
 
 	const onClick = (): void => {
